+++ atlasdb-impl-shared/src/test/java/com/palantir/atlasdb/sweep/SweepLocksTest.java
@@ -46,7 +46,8 @@ public class SweepLocksTest {
 
     @Test
     public void lockStoredInToken() throws InterruptedException {
-        when(mockLockService.lock(anyString(), any())).thenReturn(new LockRefreshToken(BigInteger.ONE, 10000000000L));
+        when(mockLockService.lock(anyString(), any()))
+                .thenReturn(new LockRefreshToken(BigInteger.ONE, 10000000000L, null));
         sweepLocks.lockOrRefresh();
 
         assertTrue(sweepLocks.haveLocks());
@@ -54,7 +55,8 @@ public class SweepLocksTest {
 
     @Test
     public void lockClearedWhenRefreshReturnsEmpty() throws InterruptedException {
-        when(mockLockService.lock(anyString(), any())).thenReturn(new LockRefreshToken(BigInteger.ONE, 10000000000L));
+        when(mockLockService.lock(anyString(), any()))
+                .thenReturn(new LockRefreshToken(BigInteger.ONE, 10000000000L, null));
         sweepLocks.lockOrRefresh();
 
         when(mockLockService.refreshLockRefreshTokens(any())).thenReturn(ImmutableSet.of());
@@ -72,7 +74,7 @@ public class SweepLocksTest {
 
     @Test
     public void lockOrRefreshCallsRefreshWhenTokenPresent() throws InterruptedException {
-        LockRefreshToken token = new LockRefreshToken(BigInteger.ONE, 10000000000L);
+        LockRefreshToken token = new LockRefreshToken(BigInteger.ONE, 10000000000L, null);
         when(mockLockService.lock(anyString(), any())).thenReturn(token);
         sweepLocks.lockOrRefresh();
         verify(mockLockService, atLeastOnce()).lock(any(), any());
@@ -84,7 +86,7 @@ public class SweepLocksTest {
 
     @Test
     public void closeUnlocksToken() throws InterruptedException {
-        LockRefreshToken token = new LockRefreshToken(BigInteger.ONE, 10000000000L);
+        LockRefreshToken token = new LockRefreshToken(BigInteger.ONE, 10000000000L, null);
         when(mockLockService.lock(anyString(), any())).thenReturn(token);
         sweepLocks.lockOrRefresh();