+++ lock-api/src/main/java/com/palantir/lock/LockRequest.java
@@ -361,6 +361,7 @@ import com.google.common.collect.Iterables;
             return this;
         }
 
+
         /**
          * Builds a {@link LockRequest} instance. After calling this method, the
          * builder object is invalidated.
@@ -410,13 +411,13 @@ import com.google.common.collect.Iterables;
         private final String creatingThreadName;
 
         SerializationProxy(LockRequest lockRequest) {
-            lockMap = lockRequest.lockMap;
-            lockTimeout = lockRequest.lockTimeout;
-            lockGroupBehavior = lockRequest.lockGroupBehavior;
-            blockingMode = lockRequest.blockingMode;
-            blockingDuration = lockRequest.blockingDuration;
-            versionId = lockRequest.versionId;
-            creatingThreadName = lockRequest.creatingThreadName;
+            this.lockMap = lockRequest.lockMap;
+            this.lockTimeout = lockRequest.lockTimeout;
+            this.lockGroupBehavior = lockRequest.lockGroupBehavior;
+            this.blockingMode = lockRequest.blockingMode;
+            this.blockingDuration = lockRequest.blockingDuration;
+            this.versionId = lockRequest.versionId;
+            this.creatingThreadName = lockRequest.creatingThreadName;
         }
 
         @JsonCreator
@@ -461,6 +462,7 @@ import com.google.common.collect.Iterables;
                 builder.withLockedInVersionId(versionId);
             }
             builder.withCreatingThreadName(creatingThreadName);
+
             return builder.build();
         }
     }