+++ docs/source/release_notes/release-notes.rst
@@ -158,6 +158,11 @@ v0.59.0
            Developers may need to check your dependencies, but no other actions should be required.
            (`Pull Request <https://github.com/palantir/atlasdb/pull/2399>`__)
 
+    *    - |userbreak| |improved|
+         - Thread can be added to Lock refreshing request. Lock service state logging will include refreshing thread as well.
+           Negative expiration times in logging should be fixed.
+           (`Pull Request <https://github.com/palantir/atlasdb/pull/2422>`__)
+
 .. <<<<------------------------------------------------------------------------------------------------------------->>>>
 
 =======